Pack_Table component refactor done
